# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single published project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
  /admin/projects:
    get:
      operationId: adminListProjects
      tags: [projects]
      summary: List all projects including drafts (admin only)
      responses:
        "200":
          description: All projects, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
    post:
      operationId: adminCreateProject
      tags: [projects]
      summary: Create a project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid project
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "409":
          description: Slug already exists
  /admin/projects/{id}:
    patch:
      operationId: adminUpdateProject
      tags: [projects]
      summary: Update a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid update
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Project not found
        "409":
          description: Slug already exists
    delete:
      operationId: adminDeleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Project not found
  /admin/projects/{id}/images:
    post:
      operationId: adminAddProjectImage
      tags: [projects]
      summary: Attach an uploaded image to a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectImageInput"
      responses:
        "201":
          description: Created image record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectImage"
        "400":
          description: Invalid image
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Project not found
  /admin/project-images/{id}:
    delete:
      operationId: adminDeleteProjectImage
      tags: [projects]
      summary: Delete a project image (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Image not found
...
  /admin/project-categories:
    post:
      operationId: adminCreateProjectCategory
      tags: [projects]
      summary: Create a project category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCategory"
        "400":
          description: Invalid category
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "409":
          description: Category already exists
  /admin/project-categories/{id}:
    delete:
      operationId: adminDeleteProjectCategory
      tags: [projects]
      summary: Delete a project category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Category not found
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload (admin only)
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
        '403':
          description: Authenticated but not an admin
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ProjectImage:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        imagePath: { type: string }
        kind: { type: string, enum: [desktop, mobile, dashboard, feature] }
        altText: { type: ["string", "null"] }
        sortOrder: { type: integer }
      required: [id, projectId, imagePath, kind, altText, sortOrder]
    ProjectImageInput:
      type: object
      properties:
        imagePath: { type: string, minLength: 1 }
        kind: { type: string, enum: [desktop, mobile, dashboard, feature] }
        altText: { type: string }
        sortOrder: { type: integer }
      required: [imagePath]
    Project:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        category: { type: string }
        shortDescription: { type: string }
        overview: { type: string }
        clientIndustry: { type: ["string", "null"] }
        challenge: { type: string }
        solution: { type: string }
        techStack:
          type: array
          items: { type: string }
        processNotes: { type: ["string", "null"] }
        results: { type: ["string", "null"] }
        lessonsLearned: { type: ["string", "null"] }
        thumbnailPath: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        published: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProjectImage"
      required:
        [
          id,
          slug,
          title,
          category,
          shortDescription,
          overview,
          clientIndustry,
          challenge,
          solution,
          techStack,
          processNotes,
          results,
          lessonsLearned,
          thumbnailPath,
          completedAt,
          published,
          seoTitle,
          seoDescription,
          createdAt,
          updatedAt,
          images,
        ]
    ProjectInput:
      type: object
      properties:
        slug: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        shortDescription: { type: string, minLength: 1 }
        overview: { type: string, minLength: 1 }
        clientIndustry: { type: ["string", "null"] }
        challenge: { type: string, minLength: 1 }
        solution: { type: string, minLength: 1 }
        techStack:
          type: array
          items: { type: string }
        processNotes: { type: ["string", "null"] }
        results: { type: ["string", "null"] }
        lessonsLearned: { type: ["string", "null"] }
        thumbnailPath: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        published: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
      required: [slug, title, category, shortDescription, overview, challenge, solution, techStack]
    ProjectUpdateInput:
      type: object
      properties:
        slug: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        shortDescription: { type: string, minLength: 1 }
        overview: { type: string, minLength: 1 }
        clientIndustry: { type: ["string", "null"] }
        challenge: { type: string, minLength: 1 }
        solution: { type: string, minLength: 1 }
        techStack:
          type: array
          items: { type: string }
        processNotes: { type: ["string", "null"] }
        results: { type: ["string", "null"] }
        lessonsLearned: { type: ["string", "null"] }
        thumbnailPath: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        published: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
    ProjectCategory:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
      required: [id, slug, name]
    ProjectCategoryInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
      required: [name]
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...