# lib/api-zod/src/generated/types/projectUpdateInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdateInput {
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  shortDescription?: string;
  /** @minLength 1 */
  overview?: string;
  /** @nullable */
  clientIndustry?: string | null;
  /** @minLength 1 */
  challenge?: string;
  /** @minLength 1 */
  solution?: string;
  techStack?: string[];
  /** @nullable */
  processNotes?: string | null;
  /** @nullable */
  results?: string | null;
  /** @nullable */
  lessonsLearned?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  completedAt?: string | null;
  published?: boolean;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
}